refactor: Split up darknet and yolo on versions
