+++ holocron/models/__init__.py
@@ -4,6 +4,9 @@ from .tridentnet import *
 from .pyconvresnet import *
 from .rexnet import *
 from .darknet import *
+from .darknetv2 import *
+from .darknetv3 import *
+from .darknetv4 import *
 from .sknet import *
 from .detection import *
 from .segmentation import *
