+++ holocron/models/detection/__init__.py
@@ -1,1 +1,3 @@
 from .yolo import *
+from .yolov2 import *
+from .yolov4 import *
