+++ holocron/models/detection/yolo.py
@@ -1,45 +1,46 @@
-# -*- coding: utf-8 -*-
-
-"""
-Personal implementation of YOLO models
-"""
-
 import sys
 import logging
 import torch
+from torch import Tensor
 import torch.nn as nn
 import torch.nn.functional as F
 from torchvision.ops.boxes import box_iou, nms
 from torchvision.ops.misc import FrozenBatchNorm2d
 from torchvision.models.utils import load_state_dict_from_url
+from typing import Dict, Optional, Any, Callable, List, Tuple, Union
 
 from ..utils import conv_sequence
-from ..darknet import DarknetBodyV1, DarknetBodyV2, DarknetBodyV4, default_cfgs as dark_cfgs
-from holocron.ops.boxes import ciou_loss
-from holocron.nn import ConcatDownsample2d, Mish, DropBlock2d, SPP, SAM
+from ..darknet import DarknetBodyV1, default_cfgs as dark_cfgs
 from holocron.nn.init import init_module
 
 
-__all__ = ['YOLOv1', 'YOLOv2', 'YOLOv4', 'yolov1', 'yolov2', 'yolov4', 'SPP', 'PAN', 'Neck']
+__all__ = ['YOLOv1', 'yolov1']
 
 
-default_cfgs = {
+default_cfgs: Dict[str, Dict[str, Any]] = {
     'yolov1': {'arch': 'YOLOv1', 'backbone': dark_cfgs['darknet24'],
-               'url': None},
-    'yolov2': {'arch': 'YOLOv2', 'backbone': dark_cfgs['darknet19'],
-               'url': None},
-    'yolov4': {'arch': 'YOLOv4', 'backbone': dark_cfgs['cspdarknet53'],
                'url': None},
 }
 
 
 class _YOLO(nn.Module):
-    def __init__(self, rpn_nms_thresh=0.7, box_score_thresh=0.05):
+    def __init__(
+        self,
+        rpn_nms_thresh: float = 0.7,
+        box_score_thresh: float = 0.05
+    ) -> None:
         super().__init__()
         self.rpn_nms_thresh = rpn_nms_thresh
         self.box_score_thresh = box_score_thresh
 
-    def _compute_losses(self, pred_boxes, pred_o, pred_scores, target, ignore_high_iou=False):
+    def _compute_losses(
+        self,
+        pred_boxes: Tensor,
+        pred_o: Tensor,
+        pred_scores: Tensor,
+        target: List[Dict[str, Tensor]],
+        ignore_high_iou=False
+    ) -> Dict[str, Tensor]:
         """Computes the detector losses as described in `"You Only Look Once: Unified, Real-Time Object Detection"
         <https://pjreddie.com/media/files/papers/yolo_1.pdf>`_
 
@@ -140,7 +141,13 @@ class _YOLO(nn.Module):
                     clf_loss=clf_loss / pred_boxes.shape[0])
 
     @staticmethod
-    def post_process(b_coords, b_o, b_scores, rpn_nms_thresh=0.7, box_score_thresh=0.05):
+    def post_process(
+        b_coords: Tensor,
+        b_o: Tensor,
+        b_scores: Tensor,
+        rpn_nms_thresh=0.7,
+        box_score_thresh=0.05
+    ) -> List[Dict[str, Tensor]]:
         """Perform final filtering to produce detections
 
         Args:
@@ -164,10 +171,9 @@ class _YOLO(nn.Module):
             # Objectness filter
             if torch.any(b_o[idx] >= 0.5):
                 coords = b_coords[idx, b_o[idx] >= 0.5]
-                scores = b_scores[idx, b_o[idx] >= 0.5].max(dim=-1)
-                labels = scores.indices
+                scores, labels = b_scores[idx, b_o[idx] >= 0.5].max(dim=-1)
                 # Multiply by the objectness
-                scores = scores.values * b_o[idx, b_o[idx] >= 0.5]
+                scores.mul_(b_o[idx, b_o[idx] >= 0.5])
 
                 # Confidence threshold
                 coords = coords[scores >= box_score_thresh]
@@ -190,9 +196,23 @@ class _YOLO(nn.Module):
 
 
 class YOLOv1(_YOLO):
-    def __init__(self, layout, num_classes=20, in_channels=3, stem_channels=64, num_anchors=2,
-                 lambda_noobj=0.5, lambda_coords=5., rpn_nms_thresh=0.7, box_score_thresh=0.05,
-                 act_layer=None, norm_layer=None, drop_layer=None, conv_layer=None, backbone_norm_layer=None):
+    def __init__(
+        self,
+        layout: List[List[int]],
+        num_classes: int = 20,
+        in_channels: int = 3,
+        stem_channels: int = 64,
+        num_anchors: int = 2,
+        lambda_noobj: float = 0.5,
+        lambda_coords: float = 5.,
+        rpn_nms_thresh: float = 0.7,
+        box_score_thresh: float = 0.05,
+        act_layer: Optional[nn.Module] = None,
+        norm_layer: Optional[Callable[[int], nn.Module]] = None,
+        drop_layer: Optional[Callable[..., nn.Module]] = None,
+        conv_layer: Optional[Callable[..., nn.Module]] = None,
+        backbone_norm_layer: Optional[Callable[[int], nn.Module]] = None
+    ) -> None:
 
         super().__init__(rpn_nms_thresh, box_score_thresh)
 
@@ -229,7 +249,7 @@ class YOLOv1(_YOLO):
         init_module(self.block4, 'leaky_relu')
         init_module(self.classifier, 'leaky_relu')
 
-    def _format_outputs(self, x):
+    def _format_outputs(self, x: Tensor) -> Tuple[Tensor, Tensor, Tensor]:
         """Formats convolutional layer output
 
         Args:
@@ -266,7 +286,7 @@ class YOLOv1(_YOLO):
 
         return b_coords, b_o, b_scores
 
-    def _forward(self, x):
+    def _forward(self, x: Tensor) -> Tensor:
 
         out = self.backbone(x)
         out = self.block4(out)
@@ -274,7 +294,11 @@ class YOLOv1(_YOLO):
 
         return out
 
-    def forward(self, x, target=None):
+    def forward(
+        self,
+        x: Tensor,
+        target: Optional[List[Dict[str, Tensor]]] = None
+    ) -> Union[Dict[str, Tensor], List[Dict[str, Tensor]]]:
         """Perform detection on an image tensor and returns either the loss dictionary in training mode
         or the list of detections in eval mode.
 
@@ -297,7 +321,7 @@ class YOLOv1(_YOLO):
 
         if self.training:
             # Update losses
-            return self._compute_losses(b_coords, b_o, b_scores, target)
+            return self._compute_losses(b_coords, b_o, b_scores, target)  # type: ignore[arg-type]
         else:
             # B * (H * W * num_anchors)
             b_coords = b_coords.view(b_coords.shape[0], -1, 4)
@@ -310,528 +334,7 @@ class YOLOv1(_YOLO):
             return self.post_process(b_coords, b_o, b_scores, self.rpn_nms_thresh, self.box_score_thresh)
 
 
-class YOLOv2(_YOLO):
-
-    def __init__(self, layout, num_classes=20, in_channels=3, stem_chanels=32, anchors=None, passthrough_ratio=8,
-                 lambda_noobj=0.5, lambda_coords=5., rpn_nms_thresh=0.7, box_score_thresh=0.05,
-                 act_layer=None, norm_layer=None, drop_layer=None, conv_layer=None, backbone_norm_layer=None):
-
-        super().__init__(rpn_nms_thresh, box_score_thresh)
-
-        if act_layer is None:
-            act_layer = nn.LeakyReLU(0.1, inplace=True)
-        if norm_layer is None:
-            norm_layer = nn.BatchNorm2d
-        if backbone_norm_layer is None:
-            backbone_norm_layer = norm_layer
-
-        # Priors computed using K-means
-        if anchors is None:
-            anchors = torch.tensor([[1.3221, 1.73145], [3.19275, 4.00944], [5.05587, 8.09892],
-                                    [9.47112, 4.84053], [11.2364, 10.0071]])
-        self.num_classes = num_classes
-
-        self.backbone = DarknetBodyV2(layout, in_channels, stem_chanels, True, act_layer,
-                                      backbone_norm_layer, drop_layer, conv_layer)
-
-        self.block5 = nn.Sequential(
-            *conv_sequence(layout[-1][0], layout[-1][0], act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(layout[-1][0], layout[-1][0], act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False))
-
-        self.passthrough_layer = nn.Sequential(*conv_sequence(layout[-2][0], layout[-2][0] // passthrough_ratio,
-                                                              act_layer, norm_layer, drop_layer, conv_layer,
-                                                              kernel_size=1, bias=False),
-                                               ConcatDownsample2d(scale_factor=2))
-
-        self.block6 = nn.Sequential(
-            *conv_sequence(layout[-1][0] + layout[-2][0] // passthrough_ratio * 2 ** 2, layout[-1][0],
-                           act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False))
-
-        # Each box has P_objectness, 4 coords, and score for each class
-        self.head = nn.Conv2d(layout[-1][0], anchors.shape[0] * (5 + num_classes), 1)
-
-        # Register losses
-        self.register_buffer('anchors', anchors)
-
-        # Loss coefficients
-        self.lambda_noobj = lambda_noobj
-        self.lambda_coords = lambda_coords
-
-        init_module(self.block5, 'leaky_relu')
-        init_module(self.passthrough_layer, 'leaky_relu')
-        init_module(self.block6, 'leaky_relu')
-        init_module(self.head, 'leaky_relu')
-
-    @property
-    def num_anchors(self):
-        return self.anchors.shape[0]
-
-    def _format_outputs(self, x):
-        """Formats convolutional layer output
-
-        Args:
-            x (torch.Tensor[N, num_anchors * (5 + num_classes), H, W]): output tensor
-
-        Returns:
-            torch.Tensor[N, H, W, num_anchors, 4]: relative coordinates in format (x, y, w, h)
-            torch.Tensor[N, H, W, num_anchors]: objectness scores
-            torch.Tensor[N, H, W, num_anchors, num_classes]: classification scores
-        """
-
-        b, _, h, w = x.shape
-        # B * C * H * W --> B * H * W * num_anchors * (5 + num_classes)
-        x = x.view(b, self.num_anchors, 5 + self.num_classes, h, w).permute(0, 3, 4, 1, 2)
-        # Cell offset
-        c_x = torch.arange(w, dtype=torch.float, device=x.device)
-        c_y = torch.arange(h, dtype=torch.float, device=x.device)
-        # Box coordinates
-        b_x = (torch.sigmoid(x[..., 0]) + c_x.view(1, 1, -1, 1)) / w
-        b_y = (torch.sigmoid(x[..., 1]) + c_y.view(1, -1, 1, 1)) / h
-        b_w = self.anchors[:, 0].view(1, 1, 1, -1) / w * torch.exp(x[..., 2])
-        b_h = self.anchors[:, 1].view(1, 1, 1, -1) / h * torch.exp(x[..., 3])
-        # B * H * W * num_anchors * 4
-        b_coords = torch.stack((b_x, b_y, b_w, b_h), dim=4)
-        # Objectness
-        b_o = torch.sigmoid(x[..., 4])
-        # Classification scores
-        b_scores = F.softmax(x[..., 5:], dim=-1)
-
-        return b_coords, b_o, b_scores
-
-    def _forward(self, x):
-
-        out, passthrough = self.backbone(x)
-        # Downsample the feature map by stacking adjacent features on the channel dimension
-        passthrough = self.passthrough_layer(passthrough)
-
-        out = self.block5(out)
-        # Stack the downsampled feature map on the channel dimension
-        out = torch.cat((passthrough, out), 1)
-        out = self.block6(out)
-
-        out = self.head(out)
-
-        return out
-
-    def forward(self, x, target=None):
-        """Perform detection on an image tensor and returns either the loss dictionary in training mode
-        or the list of detections in eval mode.
-
-        Args:
-            x (torch.Tensor[N, 3, H, W]): input image tensor
-            target (list<dict>, optional): each dict must have two keys `boxes` of type torch.Tensor[-1, 4]
-            and `labels` of type torch.Tensor[-1]
-        """
-
-        if self.training and target is None:
-            raise ValueError("`target` needs to be specified in training mode")
-
-        if isinstance(x, (list, tuple)):
-            x = torch.stack(x, dim=0)
-
-        out = self._forward(x)
-
-        # B * H * W * num_anchors
-        b_coords, b_o, b_scores = self._format_outputs(out)
-
-        if self.training:
-            # Update losses
-            return self._compute_losses(b_coords, b_o, b_scores, target)
-        else:
-            # B * (H * W * num_anchors)
-            b_coords = b_coords.view(b_coords.shape[0], -1, 4)
-            b_o = b_o.view(b_o.shape[0], -1)
-            b_scores = b_scores.reshape(b_scores.shape[0], -1, self.num_classes)
-
-            # Stack detections into a list
-            return self.post_process(b_coords, b_o, b_scores, self.rpn_nms_thresh, self.box_score_thresh)
-
-
-class PAN(nn.Module):
-    """PAN layer from `"Path Aggregation Network for Instance Segmentation" <https://arxiv.org/pdf/1803.01534.pdf>`_.
-
-    Args:
-        in_channels (int): input channels
-        act_layer (torch.nn.Module, optional): activation layer to be used
-        norm_layer (callable, optional): normalization layer
-        drop_layer (callable, optional): regularization layer
-        conv_layer (callable, optional): convolutional layer
-    """
-    def __init__(self, in_channels, act_layer=None, norm_layer=None, drop_layer=None, conv_layer=None):
-        super().__init__()
-
-        self.conv1 = nn.Sequential(*conv_sequence(in_channels, in_channels // 2,
-                                                  act_layer, norm_layer, drop_layer, conv_layer,
-                                                  kernel_size=1, bias=False))
-        self.up = nn.Upsample(scale_factor=2, mode='nearest')
-
-        self.conv2 = nn.Sequential(*conv_sequence(in_channels, in_channels // 2,
-                                                  act_layer, norm_layer, drop_layer, conv_layer,
-                                                  kernel_size=1, bias=False))
-
-        self.convs = nn.Sequential(
-            *conv_sequence(in_channels, in_channels // 2, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(in_channels // 2, in_channels, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(in_channels, in_channels // 2, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(in_channels // 2, in_channels, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(in_channels, in_channels // 2, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False))
-
-    def forward(self, x, up):
-        out = self.conv1(x)
-
-        out = torch.cat([self.conv2(up), self.up(out)], dim=1)
-
-        return self.convs(out)
-
-
-class Neck(nn.Module):
-    def __init__(self, in_planes, act_layer=None, norm_layer=None, drop_layer=None, conv_layer=None):
-        super().__init__()
-
-        self.fpn = nn.Sequential(
-            *conv_sequence(in_planes[0], in_planes[0] // 2, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(in_planes[0] // 2, in_planes[0], act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(in_planes[0], in_planes[0] // 2, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            SPP([5, 9, 13]),
-            *conv_sequence(4 * in_planes[0] // 2, in_planes[0] // 2, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(in_planes[0] // 2, in_planes[0], act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(in_planes[0], in_planes[0] // 2, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False)
-        )
-
-        self.pan1 = PAN(in_planes[1], act_layer, norm_layer, drop_layer, conv_layer)
-        self.pan2 = PAN(in_planes[2], act_layer, norm_layer, drop_layer, conv_layer)
-        init_module(self, 'leaky_relu')
-
-    def forward(self, feats):
-
-        out = self.fpn(feats[2])
-
-        aux1 = self.pan1(out, feats[1])
-        aux2 = self.pan2(aux1, feats[0])
-
-        return aux2, aux1, out
-
-
-class YoloLayer(nn.Module):
-    """Scale-specific part of YoloHead"""
-    def __init__(self, anchors, num_classes=80, scale_xy=1, iou_thresh=0.213,
-                 lambda_noobj=0.5, lambda_coords=5., rpn_nms_thresh=0.7, box_score_thresh=0.05):
-        super().__init__()
-        self.num_classes = num_classes
-        self.register_buffer('anchors', anchors)
-
-        self.rpn_nms_thresh = rpn_nms_thresh
-        self.box_score_thresh = box_score_thresh
-        self.lambda_noobj = lambda_noobj
-        self.lambda_coords = lambda_coords
-
-        # cf. https://github.com/AlexeyAB/darknet/blob/master/cfg/yolov4.cfg#L1150
-        self.scale_xy = scale_xy
-        # cf. https://github.com/AlexeyAB/darknet/blob/master/cfg/yolov4.cfg#L1151
-        self.iou_thresh = iou_thresh
-
-    def extra_repr(self):
-        return f"num_classes={self.num_classes}, scale_xy={self.scale_xy}"
-
-    def _format_outputs(self, output):
-        b, _, h, w = output.shape
-
-        # B x (num_anchors * (5 + num_classes)) x H x W --> B x H x W x num_anchors x (5 + num_classes)
-        output = output.reshape(b, len(self.anchors), 5 + self.num_classes, h, w).permute(0, 3, 4, 1, 2)
-
-        # Box center
-        c_x = torch.arange(w, dtype=torch.float32, device=output.device).view(1, 1, -1, 1)
-        c_y = torch.arange(h, dtype=torch.float32, device=output.device).view(1, -1, 1, 1)
-
-        b_xy = self.scale_xy * torch.sigmoid(output[..., :2]) - 0.5 * (self.scale_xy - 1)
-        b_xy[..., 0] += c_x
-        b_xy[..., 1] += c_y
-        b_xy[..., 0] /= w
-        b_xy[..., 1] /= h
-
-        # Box dimension
-        b_wh = torch.exp(output[..., 2:4]) * self.anchors.view(1, 1, 1, -1, 2)
-
-        top_left = b_xy - 0.5 * b_wh
-        bot_right = top_left + b_wh
-        boxes = torch.cat((top_left, bot_right), dim=-1)
-
-        # Objectness
-        b_o = torch.sigmoid(output[..., 4])
-        # Classification scores
-        b_scores = torch.sigmoid(output[..., 5:])
-
-        return boxes, b_o, b_scores
-
-    @staticmethod
-    def post_process(boxes, b_o, b_scores, rpn_nms_thresh=0.7, box_score_thresh=0.05):
-
-        boxes = boxes.clamp_(0, 1)
-        detections = []
-        for idx in range(b_o.shape[0]):
-
-            coords = torch.zeros((0, 4), dtype=torch.float32, device=b_o.device)
-            scores = torch.zeros(0, dtype=torch.float32, device=b_o.device)
-            labels = torch.zeros(0, dtype=torch.long, device=b_o.device)
-
-            # Objectness filter
-            if torch.any(b_o[idx] >= 0.5):
-                coords = boxes[idx, b_o[idx] >= 0.5]
-                scores = b_scores[idx, b_o[idx] >= 0.5].max(dim=-1)
-                labels = scores.indices
-                # Multiply by the objectness
-                scores = scores.values * b_o[idx, b_o[idx] >= 0.5]
-
-                # Confidence threshold
-                coords = coords[scores >= box_score_thresh]
-                labels = labels[scores >= box_score_thresh]
-                scores = scores[scores >= box_score_thresh]
-                coords = coords.clamp_(0, 1)
-                # NMS
-                kept_idxs = nms(coords, scores, iou_threshold=rpn_nms_thresh)
-                coords = coords[kept_idxs]
-                scores = scores[kept_idxs]
-                labels = labels[kept_idxs]
-
-            detections.append(dict(boxes=coords, scores=scores, labels=labels))
-
-        return detections
-
-    def _build_targets(self, pred_boxes, b_o, b_scores, target):
-
-        b, h, w, num_anchors = b_o.shape
-
-        # Target formatting
-        target_o = torch.zeros((b, h, w, num_anchors), device=b_o.device)
-        target_scores = torch.zeros((b, h, w, num_anchors, self.num_classes), device=b_o.device)
-        obj_mask = torch.zeros((b, h, w, num_anchors), dtype=torch.bool, device=b_o.device)
-        noobj_mask = torch.ones((b, h, w, num_anchors), dtype=torch.bool, device=b_o.device)
-
-        gt_boxes = [t['boxes'] for t in target]
-        gt_labels = [t['labels'] for t in target]
-
-        # GT coords --> left, top, width, height
-        _boxes = torch.cat(gt_boxes, dim=0)
-        gt_centers = _boxes[..., [0, 2, 1, 3]].view(-1, 2, 2).mean(dim=-1)
-        gt_centers[:, 0] *= w
-        gt_centers[:, 1] *= h
-        gt_centers = gt_centers.to(dtype=torch.long)
-
-        target_selection = torch.tensor([_idx for _idx, _boxes in enumerate(gt_boxes) for _ in range(_boxes.shape[0])],
-                                        dtype=torch.long, device=b_o.device)
-        if target_selection.shape[0] > 0:
-
-            # Anchors IoU
-            gt_wh = _boxes[:, 2:] - _boxes[:, :2]
-            anchor_idxs = box_iou(torch.cat((-gt_wh, gt_wh), dim=-1),
-                                  torch.cat((-self.anchors, self.anchors), dim=-1)).argmax(dim=1)
-
-            # Assign boxes
-            obj_mask[target_selection, gt_centers[:, 1], gt_centers[:, 0], anchor_idxs] = True
-            noobj_mask[target_selection, gt_centers[:, 1], gt_centers[:, 0], anchor_idxs] = False
-            # B * cells * predictors * info
-            for idx in range(b):
-                if gt_boxes[idx].shape[0] > 0:
-                    # IoU with cells that enclose the GT centers
-                    gt_ious, gt_idxs = box_iou(pred_boxes[idx, obj_mask[idx]], gt_boxes[idx]).max(dim=1)
-                    # Objectness target
-                    target_o[idx, obj_mask[idx]] = gt_ious
-                    # Classification target
-                    target_scores[idx, obj_mask[idx], gt_labels[idx][gt_idxs]] = 1.
-
-        return target_o, target_scores, obj_mask, noobj_mask
-
-    def _compute_losses(self, pred_boxes, b_o, b_scores, target, ignore_high_iou=False):
-
-        target_o, target_scores, obj_mask, noobj_mask = self._build_targets(pred_boxes, b_o, b_scores, target)
-
-        # Bbox regression
-        bbox_loss = torch.zeros(1, device=b_o.device)
-        for idx, _target in enumerate(target):
-            if _target['boxes'].shape[0] > 0 and torch.any(obj_mask[idx]):
-                bbox_loss += ciou_loss(pred_boxes[idx, obj_mask[idx]], _target['boxes']).min(dim=1).values.sum()
-
-        return dict(obj_loss=F.mse_loss(b_o[obj_mask], target_o[obj_mask], reduction='sum'),
-                    noobj_loss=self.lambda_noobj * b_o[noobj_mask].pow(2).sum(),
-                    bbox_loss=self.lambda_coords * bbox_loss,
-                    clf_loss=F.binary_cross_entropy(b_scores[obj_mask], target_scores[obj_mask], reduction='sum'))
-
-    def forward(self, output, target=None):
-        """Perform detection on an image tensor and returns either the loss dictionary in training mode
-        or the list of detections in eval mode.
-
-        Args:
-            x (torch.Tensor[N, 3, H, W]): input image tensor
-            target (list<dict>, optional): each dict must have two keys `boxes` of type torch.Tensor[-1, 4]
-            and `labels` of type torch.Tensor[-1]
-        """
-
-        if self.training and target is None:
-            raise ValueError("`target` needs to be specified in training mode")
-
-        pred_boxes, b_o, b_scores = self._format_outputs(output)
-
-        if self.training:
-            return self._compute_losses(pred_boxes, b_o, b_scores, target)
-        else:
-            # cf. https://github.com/Tianxiaomo/pytorch-YOLOv4/blob/master/tool/yolo_layer.py#L117
-            return self.post_process(pred_boxes, b_o, b_scores, self.rpn_nms_thresh, self.box_score_thresh)
-
-
-class Yolov4Head(nn.Module):
-    def __init__(self, num_classes=80, anchors=None,
-                 act_layer=None, norm_layer=None, drop_layer=None, conv_layer=None):
-
-        # cf. https://github.com/AlexeyAB/darknet/blob/master/cfg/yolov4.cfg#L1143
-        if anchors is None:
-            anchors = torch.tensor([[[12, 16], [19, 36], [40, 28]],
-                                   [[36, 75], [76, 55], [72, 146]],
-                                   [[142, 110], [192, 243], [459, 401]]], dtype=torch.float32) / 608
-        elif not isinstance(anchors, torch.Tensor):
-            anchors = torch.tensor(anchors, dtype=torch.float32)
-
-        if anchors.shape[0] != 3:
-            raise AssertionError(f"The number of anchors is expected to be 3. received: {anchors.shape[0]}")
-
-        super().__init__()
-
-        self.head1 = nn.Sequential(
-            *conv_sequence(128, 256, act_layer, norm_layer, None, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(256, (5 + num_classes) * 3, None, None, None, conv_layer,
-                           kernel_size=1, bias=True))
-
-        self.yolo1 = YoloLayer(anchors[0], num_classes=num_classes, scale_xy=1.2)
-
-        self.pre_head2 = nn.Sequential(*conv_sequence(128, 256, act_layer, norm_layer, drop_layer, conv_layer,
-                                                      kernel_size=3, padding=1, stride=2, bias=False))
-        self.head2_1 = nn.Sequential(
-            *conv_sequence(512, 256, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(256, 512, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(512, 256, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(256, 512, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(512, 256, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False))
-        self.head2_2 = nn.Sequential(
-            *conv_sequence(256, 512, act_layer, norm_layer, None, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(512, (5 + num_classes) * 3, None, None, None, conv_layer,
-                           kernel_size=1, bias=True))
-
-        self.yolo2 = YoloLayer(anchors[1], num_classes=num_classes, scale_xy=1.1)
-
-        self.pre_head3 = nn.Sequential(*conv_sequence(256, 512, act_layer, norm_layer, drop_layer, conv_layer,
-                                                      kernel_size=3, padding=1, stride=2, bias=False))
-        self.head3 = nn.Sequential(
-            *conv_sequence(1024, 512, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(512, 1024, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(1024, 512, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(512, 1024, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(1024, 512, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=1, bias=False),
-            *conv_sequence(512, 1024, act_layer, norm_layer, drop_layer, conv_layer,
-                           kernel_size=3, padding=1, bias=False),
-            *conv_sequence(1024, (5 + num_classes) * 3, None, None, None, conv_layer,
-                           kernel_size=1, bias=True))
-
-        self.yolo3 = YoloLayer(anchors[2], num_classes=num_classes, scale_xy=1.05)
-        init_module(self, 'leaky_relu')
-        # Zero init
-        self.head1[-1].weight.data.zero_()
-        self.head1[-1].bias.data.zero_()
-        self.head2_2[-1].weight.data.zero_()
-        self.head2_2[-1].bias.data.zero_()
-        self.head3[-1].weight.data.zero_()
-        self.head3[-1].bias.data.zero_()
-
-    def forward(self, feats, target=None):
-        o1 = self.head1(feats[0])
-
-        h2 = self.pre_head2(feats[0])
-        h2 = torch.cat([h2, feats[1]], dim=1)
-        h2 = self.head2_1(h2)
-        o2 = self.head2_2(h2)
-
-        h3 = self.pre_head3(h2)
-        h3 = torch.cat([h3, feats[2]], dim=1)
-        o3 = self.head3(h3)
-
-        # YOLO output
-        y1 = self.yolo1(o1, target)
-        y2 = self.yolo2(o2, target)
-        y3 = self.yolo3(o3, target)
-
-        if not self.training:
-
-            detections = [dict(boxes=torch.cat((det1['boxes'], det2['boxes'], det3['boxes']), dim=0),
-                               scores=torch.cat((det1['scores'], det2['scores'], det3['scores']), dim=0),
-                               labels=torch.cat((det1['labels'], det2['labels'], det3['labels']), dim=0))
-                          for det1, det2, det3 in zip(y1, y2, y3)]
-            return detections
-
-        else:
-
-            return {k: y1[k] + y2[k] + y3[k] for k in y1.keys()}
-
-
-class YOLOv4(nn.Module):
-    def __init__(self, layout, num_classes=80, in_channels=3, stem_channels=32, anchors=None,
-                 act_layer=None, norm_layer=None, drop_layer=None, conv_layer=None, backbone_norm_layer=None):
-        super().__init__()
-
-        if act_layer is None:
-            act_layer = nn.LeakyReLU(0.1, inplace=True)
-        if norm_layer is None:
-            norm_layer = nn.BatchNorm2d
-        if backbone_norm_layer is None:
-            backbone_norm_layer = norm_layer
-
-        # backbone
-        self.backbone = DarknetBodyV4(layout, in_channels, stem_channels, 3, Mish(),
-                                      backbone_norm_layer, drop_layer, conv_layer)
-        # neck
-        self.neck = Neck([1024, 512, 256], act_layer, norm_layer, drop_layer, conv_layer)
-        # head
-        self.head = Yolov4Head(num_classes, anchors, act_layer, norm_layer, drop_layer, conv_layer)
-
-        init_module(self.neck, 'leaky_relu')
-        init_module(self.head, 'leaky_relu')
-
-    def forward(self, x, target=None):
-
-        if not isinstance(x, torch.Tensor):
-            x = torch.stack(x, dim=0)
-
-        out = self.backbone(x)
-
-        x20, x13, x6 = self.neck(out)
-
-        return self.head((x20, x13, x6), target)
-
-
-def _yolo(arch, pretrained, progress, pretrained_backbone, **kwargs):
+def _yolo(arch: str, pretrained: bool, progress: bool, pretrained_backbone: bool, **kwargs: Any) -> _YOLO:
 
     if pretrained:
         pretrained_backbone = False
@@ -862,7 +365,7 @@ def _yolo(arch, pretrained, progress, pretrained_backbone, **kwargs):
     return model
 
 
-def yolov1(pretrained=False, progress=True, pretrained_backbone=True, **kwargs):
+def yolov1(pretrained: bool = False, progress: bool = True, pretrained_backbone: bool = True, **kwargs: Any) -> YOLOv1:
     """YOLO model from
     `"You Only Look Once: Unified, Real-Time Object Detection" <https://pjreddie.com/media/files/papers/yolo_1.pdf>`_.
 
@@ -928,67 +431,4 @@ def yolov1(pretrained=False, progress=True, pretrained_backbone=True, **kwargs):
         torch.nn.Module: detection module
     """
 
-    return _yolo('yolov1', pretrained, progress, pretrained_backbone, **kwargs)
-
-
-def yolov2(pretrained=False, progress=True, pretrained_backbone=True, **kwargs):
-    """YOLOv2 model from
-    `"YOLO9000: Better, Faster, Stronger" <https://pjreddie.com/media/files/papers/YOLO9000.pdf>`_.
-
-    YOLOv2 improves upon YOLO by raising the number of boxes predicted by grid cell (default: 5), introducing
-    bounding box priors and predicting class scores for each anchor box in the grid cell.
-
-    For training, YOLOv2 uses the same multi-part loss as YOLO apart from its classification loss:
-
-    .. math::
-        \\mathcal{L}_{classification} = \\sum\\limits_{i=0}^{S^2}  \\sum\\limits_{j=0}^{B}
-        \\mathbb{1}_{ij}^{obj} \\sum\\limits_{c \\in classes}
-        (p_{ij}(c) - \\hat{p}_{ij}(c))^2
-
-    where :math:`S` is size of the output feature map (13 for an input size :math:`(416, 416)`),
-    :math:`B` is the number of anchor boxes per grid cell (default: 5),
-    :math:`\\mathbb{1}_{ij}^{obj}` equals to 1 if a GT center falls inside the i-th grid cell and among the
-    anchor boxes of that cell, has the highest IoU with the j-th box else 0,
-    :math:`p_{ij}(c)` equals 1 if the assigned ground truth to the j-th anchor box of the i-th cell is classified
-    as class :math:`c`,
-    and :math:`\\hat{p}_{ij}(c)` is the predicted probability of class :math:`c` for the j-th anchor box
-    in the i-th cell.
-
-    Args:
-        pretrained (bool, optional): If True, returns a model pre-trained on ImageNet
-        progress (bool, optional): If True, displays a progress bar of the download to stderr
-        pretrained_backbone (bool, optional): If True, backbone parameters will have been pretrained on Imagenette
-
-    Returns:
-        torch.nn.Module: detection module
-    """
-
-    if pretrained_backbone:
-        kwargs['backbone_norm_layer'] = FrozenBatchNorm2d
-
-    return _yolo('yolov2', pretrained, progress, pretrained_backbone, **kwargs)
-
-
-def yolov4(pretrained=False, progress=True, pretrained_backbone=True, **kwargs):
-    """YOLOv4 model from
-    `"YOLOv4: Optimal Speed and Accuracy of Object Detection" <https://arxiv.org/pdf/2004.10934.pdf>`_.
-
-    YOLOv4 is an improvement on YOLOv3 that includes many changes including: the usage of `DropBlock
-    <https://arxiv.org/pdf/1810.12890.pdf>`_ regularization, `Mish <https://arxiv.org/pdf/1908.08681.pdf>`_ activation,
-    `CSP <https://arxiv.org/pdf/2004.10934.pdf>`_ and `SAM <https://arxiv.org/pdf/1807.06521.pdf>`_ in the
-    backbone, `SPP <https://arxiv.org/pdf/1406.4729.pdf>`_ and `PAN <https://arxiv.org/pdf/1803.01534.pdf>`_ in the
-    neck.
-
-    Args:
-        pretrained (bool, optional): If True, returns a model pre-trained on ImageNet
-        progress (bool, optional): If True, displays a progress bar of the download to stderr
-        pretrained_backbone (bool, optional): If True, backbone parameters will have been pretrained on Imagenette
-
-    Returns:
-        torch.nn.Module: detection module
-    """
-
-    if pretrained_backbone:
-        kwargs['backbone_norm_layer'] = FrozenBatchNorm2d
-
-    return _yolo('yolov4', pretrained, progress, pretrained_backbone, **kwargs)
+    return _yolo('yolov1', pretrained, progress, pretrained_backbone, **kwargs)  # type: ignore[return-value]
